+++ server/src/main.rs
@@ -1,48 +1,20 @@
-use async_graphql::{http::GraphiQLSource, EmptyMutation, EmptySubscription, Object, Schema};
-use async_graphql_axum::{GraphQLRequest, GraphQLResponse};
-use axum::{
-    extract::Extension,
-    response::{self, IntoResponse},
-    routing::{get, post},
-    Router,
-};
 use std::net::SocketAddr;
 
-struct Query;
-
-#[Object]
-impl Query {
-    async fn value(&self) -> i32 {
-        100
-    }
-}
-
-type AppSchema = Schema<Query, EmptyMutation, EmptySubscription>;
-
-async fn graphql_handler(schema: Extension<AppSchema>, req: GraphQLRequest) -> GraphQLResponse {
-    schema.execute(req.into_inner()).await.into()
-}
-
-async fn graphiql() -> impl IntoResponse {
-    response::Html(GraphiQLSource::build().endpoint("/graphql").finish())
-}
+mod handlers;
+mod router;
+mod schema;
 
 #[tokio::main]
 async fn main() {
     tracing_subscriber::fmt::init();
 
-    let schema = Schema::build(Query, EmptyMutation, EmptySubscription).finish();
-
-    let app = Router::new()
-        .route("/", get(graphiql))
-        .route("/graphql", post(graphql_handler))
-        .layer(Extension(schema));
+    let router = router::create_router();
 
     let socket_address = SocketAddr::from(([127, 0, 0, 1], 3000));
-    tracing::info!("listening on {}", socket_address);
+    tracing::info!("Server running on {}", socket_address);
 
     axum::Server::bind(&socket_address)
-        .serve(app.into_make_service())
+        .serve(router.into_make_service())
         .await
         .unwrap();
 }
